backend/requirements.txt
fastapi
uvicorn[standard]

ultralytics

opencv-python-headless
numpy
Pillow

python-multipart

# optionnel mais recommandé
scipy